🔥 refactor(home_ui.dart): remove unused imports and variables This commit removes unused imports and variables in the home_ui.dart file. The import for the course_selector_ui.dart file is removed as it is not used in the file. The imports for the active_students.dart and home_course.dart files are removed as they are not used in the file. The studentsData variable is also removed as it is not used in the file.
✨ feat(home_ui.dart): add MasterLevelCubit to fetch master level data
This commit adds the MasterLevelCubit to the home_ui.dart file to fetch the master level data. The getMasterLevel method is called in the initState method to fetch the data when the page is loaded. The data is then displayed in the UI.

🔨 refactor(home_ui.dart): extract TileMasterLevel widget
The TileMasterLevel widget is extracted from the HomeItems widget to improve readability and maintainability of the code. The TileMasterLevel widget is responsible for rendering the individual tiles in the GridView.

🚀 feat(providers.dart): add MasterLevelCubit to AppProviders
The MasterLevelCubit is added to the AppProviders, which provides the necessary dependencies for the MasterLevelCubit to function. This allows the MasterLevelCubit to be used throughout the application.