🐛 fix(report_cubit.dart): fix typo in ReportDataSource variable name 🎨 style(home_ui_tile_master_loading.dart): remove unused import and reformat code The variable name in ReportCubit was misspelled as ReportDatasource instead of ReportDataSource. This commit fixes the typo to ensure that the correct class is being used. In HomeUiTileMasterLoading, an unused import was removed and the code was reformatted to improve readability.
✨ feat(reporting_ui_report_tile.dart): add ReportTile widget
This commit adds a new widget called ReportTile, which is used to display a report tile in the reporting UI. The widget takes in four required parameters: index, reportName, state, and reportImage. It displays the report name and an image associated with the report. If the report is locked, it displays a lock icon, and if it is unlocked, it displays the report image. The widget also has an onTap function that is currently commented out.

🚀 chore(reporting_ui.dart, providers.dart): add ReportCubit and use it in ReportingPage
✨ feat(reporting_ui.dart): add ReportTileLoading widget
The ReportCubit is added to the providers list in providers.dart and used in ReportingPage to fetch the report list data. The ReportTileLoading widget is added to show a loading state while the report list data is being fetched.