🚀 chore(reading_section_ui.dart): add top_snackbar_flutter package The top_snackbar_flutter package is added to the dependencies to show custom snackbars. The NotFound widget is replaced with a custom widget to improve the user experience. The success snackbar is shown when the user finishes reading a material.
🚀 chore(reporting_ui_report_tile.dart): add top_snackbar_flutter package
The top_snackbar_flutter package is added to the dependencies to show custom snackbars. The error snackbar is shown when the user tries to access a report that is not yet available.

🚀 chore(reporting_detail_ui.dart): add textAlign to Text widget
The textAlign property is added to the Text widget to align the text to the right.

🚀 chore(dialog_box.dart): add btnOkColor to SuccessDialog
The btnOkColor property is added to the SuccessDialog widget to change the color of the OK button.