🎨 style(reading_section_ui.dart): add loader overlay and lottie animation 🐛 fix(reading_section_ui.dart): fix disabledRoute property value The commit adds a loader overlay with a lottie animation to the ReadingSection UI. The loader overlay is shown while the data is being fetched. The disabledRoute property value is fixed to true to disable the back button on the app bar.
🐛 fix(signin_ui_body_login.dart): add loader overlay when logging in
✨ feat(signin_ui.dart): add loader overlay when logging in
The loader overlay is added to indicate that the application is processing the login request. The loader overlay is shown when the login button is pressed and hidden when the login request is completed. The loader overlay package is added as a dependency to the project.

🐛 fix(appbar_style.dart): replace dialog with top snackbar when user tries to go back
The dialog box is replaced with a top snackbar to inform the user that they cannot go back to the previous page when they are still working on a task. The top snackbar is more user-friendly and less intrusive than the dialog box. The top_snackbar_flutter package is added as a dependency to the project.