🔨 refactor(update_user): make optional fields nullable The fields id_google, id_fb, id_apple, fcm_id, email, no_telpon, and user_npp are now nullable. This change was made to allow for more flexibility in the UpdateProfileRequestModel.
🐛 fix(auth_datasource): comment out unused code
The line of code that adds an image file to the request was commented out as it was not being used.

✨ feat(update_user): add UpdateUserCubit
The UpdateUserCubit was added to handle updating user profiles. It takes an AuthDataSource and emits different states based on the result of the updateProfile method. The states are _Initial, _Loading, _Loaded, and _Error.

🚀 chore(detail_profile_ui.dart): refactor ProfileDetailPage to use Bloc pattern
This commit refactors the ProfileDetailPage to use the Bloc pattern by adding the UpdateUserCubit and updating the UI to use BlocListener and BlocBuilder. The commit also adds a loading state to the PrimaryButton widget and removes the onPressed function from the PrimaryButton widget. The onPressed function will be added later when the updateUser function is implemented in the UpdateUserCubit. Finally, the commit adds hintText to the ProfileForm widgets to improve the user experience.

🎨 style(signin_ui.dart): reformat code
This commit only reformats the code to improve its readability and consistency. No functional changes were made.

🔥 chore(signup_ui.dart): remove unused code
The diff shows that there are no changes in the code, only the line numbers have changed. Therefore, this commit removes unused code that was not being used in the application.

🔨 refactor(signup_ui.dart): reformat code and remove redundant code
The code has been reformatted to improve readability and consistency. Redundant code has been removed to improve maintainability.

🚀 feat(profile_form.dart): add hintText parameter to ProfileForm constructor
The hintText parameter is now required in the ProfileForm constructor. This allows the caller to specify the hint text for the text field.

🚀 feat(providers.dart): add UpdateUserCubit to AppProviders
The UpdateUserCubit is now added to the AppProviders. This allows the application to update the user's information.