🎨 style(detail_profile_ui.dart): format code This commit only formats the code in the detail_profile_ui.dart file to improve readability and consistency. No functional changes were made.
🎨 style(detail_profile_ui.dart): refactor ProfileDetailPage layout
✨ feat(detail_profile_ui.dart): add scrollable SingleChildScrollView to ProfileDetailPage
The ProfileDetailPage layout has been refactored to improve the user experience. The layout now includes a SingleChildScrollView widget to allow for scrolling when the keyboard is open. This ensures that all form fields are accessible to the user.

🚀 chore(thunderclient.json): update thunderclient.json test data
🔧 fix(thunderclient.json): fix img_file value in thunderclient.json test data
The test data in thunderclient.json has been updated to reflect the latest changes in the API. The token_api value has been updated to a new UUID. The user_id and full_name values have been updated to reflect new test data. The img_file value has been fixed to point to the correct file path.